+++ lib/keela/strategies/methods.rb
@@ -7,7 +7,7 @@ module Keela
         "methods"
       end
 
-      def definition_file_pattern
+      def default_definition_file_pattern
         %r{app/helpers|app/models}
       end